+++ clang/test/SemaTemplate/lambda-capture-pack.cpp
@@ -75,6 +75,27 @@ template <class = void> void f() {
       using T = identity<Is>::type;
     }(), ...);
   }(1, 2);
+
+  // https://github.com/llvm/llvm-project/issues/18873
+  [](auto ...y) {
+    ([y] { }, ...);
+  }();
+
+  [](auto ...x) {
+    ([&](auto ...y) {
+      // FIXME: This now hits the assertion with `PackIdx != -1 && "found declaration pack but not pack expanding"'
+      // in Sema::FindInstantiatedDecl.
+      // This is because the captured variable x has been expanded while transforming
+      // the outermost lambda call, but the expansion then gets holded off while transforming
+      // the folded expression. Then we would hit the assertion when instantiating the captured variable
+      // in TransformLambdaExpr.
+      // I think this is supposed to be ill-formed, but GCC and MSVC currently accept this.
+      // If x gets expanded with non-empty arguments, then GCC and MSVC will reject it - we probably
+      // miss a diagnostic for it.
+      // ([x, y] { }, ...);
+      ([x..., y] { }, ...);
+    })();
+  }();
 }
 
 template void f();